+++ Runtime/MonoHierarchyIconDrawer.cs
@@ -11,11 +11,14 @@ namespace OcUtility
 #endif
     {
 #if UNITY_EDITOR
+        public Texture2D IconTexture => iconTexture;
         public string IconPath => iconPath;
         public int DistanceToText => distanceToText;
         public Color IconTint => iconTint;
-        
-        [FilePath]public string iconPath = "EditorComment Icon";
+
+        public Texture2D iconTexture;
+        [ShowIf("iconTexture", null)][FilePath]public string iconPath = "EditorComment Icon";
+
         public int distanceToText = 15;
         public Color iconTint = Color.white;
 #endif